ASDF

+++ user-profile/src/main/java/com/microservice/erp/domain/dto/SaRoleDto.java
@@ -4,9 +4,11 @@ import com.microservice.erp.domain.helper.EnumRole;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Setter
 @Getter
 public class SaRoleDto {
-    private Integer roleId;
+    private BigInteger roleId;
     private EnumRole roleName;
 }
